\name{CCIterate}
\alias{CCIterate}
\title{Internal Function of Iterations of ConjointChecks}
\description{
Internal function; should not be used directly.
}
\usage{
CCIterate(nIter, old, old_ll, single, burn, N, n)
}
\arguments{
\item{nIter}{Number of iterations.}
\item{old}{Numeric matrix.}
\item{old_ll}{Numeric matrix.}
\item{single}{Should do single cancellation.}
\item{burn}{Number of initial values to remove.}
\item{N}{Integer matrix.}
\item{n}{Numeric matrix.}
}
\references{
Perline, R., Wright, B. D., & Wainer, H. (1979). The Rasch model as
additive conjoint measurement. Applied Psychological Measurement, 3(2), 237-255.
}
\author{Ben Domingue \email{ben.domingue@gmail.com}} 
\examples{
### INTERNAL FUNCTION ###
}

