\name{ConjointChecs-package}
\alias{ConjointChecks-package}
\docType{package}
\title{ConjointChecks: A package to check the cancellation axioms of conjoint measurement.}
\description{
Implementation of a procedure (Domingue, 2012; see also Karabatsos, 2001 and  Kyngdon, 2011) to test the single and double cancellation axioms of conjoint measure in data that is dichotomously coded and measured with error.
}


\author{Ben Domingue \email{ben.domingue@gmail.com}} 

\references{
Domingue, B. (2012). Evaluating the Equal-Interval Hypothesis with Test
Score Scales. Doctoral Dissertation, University of Colorado Boulder, May
2012.

Karabatsos, G. (2001). The rasch model, additive conjoint measurement, and new models of
probabilistic measurement theory. Journal of Applied Measurement, 2(4), 389-423.

Kyngdon, A. (2011). Plausible measurement analogies to some psychometric models of test per-
formance. British Journal of Mathematical and Statistical Psychology, 64(3), 478-497.

Perline, R., Wright, B. D., & Wainer, H. (1979). The Rasch model as
additive conjoint measurement. Applied Psychological Measurement, 3(2), 237-255.
}

\keyword{package}

