\name{list.null-class}
\docType{class}
\alias{list.null-class}
\title{Class "list.null"}
\description{The formal S4 class for list.null. This class contains a null list.}
\section{Objects from the Class}{
Object used internally. 
}
\details{
  Objects of class are used internally.
}
\author{ Ben Domingue \email{ben.domingue@gmail.com} }
\seealso{
\code{\link{ConjointChecks}}, \code{\link{summary.checks}}, \code{\link{plot.checks}}
}
\keyword{classes}
