% File src/library/base/man/load.Rd
\name{SummarizeChecks}
\alias{SummarizeChecks}
\title{Summarize checks produced by \code{ConjointChecks()}.}
\description{
Takes output from \code{ConjointChecks()} and produces a matrix showing
the percentage of reported violations at each cell.
}
\usage{
summarize_checks(chain) 
}
\arguments{
\item{chain}{Object returned by \code{ConjointChecks}.}
}
\examples{
}
\keyword{file}

